+++ packages/epub_view/example/lib/main.dart
@@ -2,12 +2,57 @@ import 'dart:typed_data';
 import 'dart:ui';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart' show rootBundle;
+import 'package:flutter/services.dart'
+    show SystemChrome, SystemUiOverlayStyle, rootBundle;
 import 'package:epub_view/epub_view.dart';
 
 void main() => runApp(MyApp());
 
-class MyApp extends StatelessWidget {
+class MyApp extends StatefulWidget {
+  @override
+  _MyAppState createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
+  @override
+  void initState() {
+    WidgetsBinding.instance.addObserver(this);
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    WidgetsBinding.instance.removeObserver(this);
+    super.dispose();
+  }
+
+  @override
+  void didChangePlatformBrightness() {
+    _setSystemUIOverlayStyle();
+  }
+
+  Brightness get platformBrightness =>
+      MediaQueryData.fromWindow(WidgetsBinding.instance.window)
+          .platformBrightness;
+
+  void _setSystemUIOverlayStyle() {
+    if (platformBrightness == Brightness.light) {
+      SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
+        statusBarBrightness: Brightness.light,
+        statusBarIconBrightness: Brightness.dark,
+        systemNavigationBarColor: Colors.grey[50],
+        systemNavigationBarIconBrightness: Brightness.dark,
+      ));
+    } else {
+      SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
+        statusBarBrightness: Brightness.dark,
+        statusBarIconBrightness: Brightness.light,
+        systemNavigationBarColor: Colors.grey[850],
+        systemNavigationBarIconBrightness: Brightness.light,
+      ));
+    }
+  }
+
   @override
   Widget build(BuildContext context) => MaterialApp(
         title: 'Epub demo',
@@ -19,6 +64,7 @@ class MyApp extends StatelessWidget {
           primarySwatch: Colors.blue,
           brightness: Brightness.dark,
         ),
+        debugShowCheckedModeBanner: false,
         home: MyHomePage(),
       );
 }
@@ -38,13 +84,20 @@ class _MyHomePageState extends State<MyHomePage> {
     final loadedBook =
         _loadFromAssets('assets/New-Findings-on-Shirdi-Sai-Baba.epub');
     _epubReaderController = EpubController(
-      data: loadedBook,
+      document: EpubReader.readBook(loadedBook),
+      //  document: EpubReader,
       // epubCfi:
       //     'epubcfi(/6/26[id4]!/4/2/2[id4]/22)', // book.epub Chapter 3 paragraph 10
       // epubCfi:
       //     'epubcfi(/6/6[chapter-2]!/4/2/1612)', // book_2.epub Chapter 16 paragraph 3
     );
     super.initState();
+  }
+
+  @override
+  void dispose() {
+    _epubReaderController.dispose();
+    super.dispose();
   }
 
   Future<Uint8List> _loadFromAssets(String assetName) async {